requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.8.3
protobuf>=4.25.0